Add lane assignment module to tournament manager
Refactor the codebase to introduce a new module for lane assignments in the `BowlingMegabucks.TournamentManager` namespace. This includes the implementation of the `GenerateCross` abstract class and its concrete implementations, `SameSkip` and `StaggeredSkip`, for managing lane assignment logic.

Introduce the `LaneAvailability` class to generate lane assignments based on specified rules. Define interfaces such as `IGenerate`, `ILaneAvailability`, and `IViewModel` to standardize interactions.

Update the `Presenter` class to handle lane assignments, including loading, updating, and generating recap sheets. Set up dependency injection in `LaneAssignmentsExtensions` for new services and adapters.

Modify the `RecapSheetViewModel` to include properties for bowler details and lane assignments, enhancing the structure of recap data. Overall, these changes improve the functionality, maintainability, and readability of the tournament manager.
