+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/GenerateCross.cs
