+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/GenerateCrossFactory.cs
