+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/GenerateLaneAvailability.cs
