+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/IGenerateLaneAssignments.cs
