+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/LaneAssignmentsExtensions.cs
