+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/Retrieve/RetrieveLaneAssignmentsAdapter.cs
