+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/SameSkipLaneAssignments.cs
