+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/StaggeredSkipLaneAssignments.cs
