+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/LaneAssignments/Update/UpdateLaneAssignmentsAdapter.cs
