+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Registrations/Delete/DeleteRegistrationAdapter.cs
