+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Registrations/Update/UpdateRegistrationAdapter.cs
